feat: add session check in every controller and fix page redirect

+++ src/provider/schedule/schedule.controller.ts
@@ -1,6 +1,7 @@
 //import packages
-import { Controller, Post, Get, Patch, Delete, Body, UseFilters, BadRequestException, Logger } from '@nestjs/common';
+import { Controller, Post, Get, Patch, Delete, Body, UseFilters, BadRequestException, Res, Session } from '@nestjs/common';
 import { ApiOperation, ApiTags } from '@nestjs/swagger';
+import { Response } from 'express';
 
 //import constants
 import { CONTROLLER } from './schedule.constants';
@@ -13,6 +14,7 @@ import { ScheduleSetup } from 'src/model/postgre/scheduleSetup/scheduleSetup.ent
 //import service
 import { ScheduleService } from './schedule.service';
 import { LoggerService } from 'src/common/logger/logger.service';
+import { HttpService } from 'src/common/http/http.service';
 
 const {
     API_TAGS,       //tag for Swagger UI
@@ -21,6 +23,7 @@ const {
     READALL_ROUTE,  //readAll
     UPDATE_ROUTE,   //update
     DELETE_ROUTE,   //delete
+    REDIRECT_ROUTES,//
 } = CONTROLLER;
 
 @ApiTags(API_TAGS)
@@ -29,17 +32,23 @@ const {
 export class ScheduleController {
     constructor(
         private readonly scheduleService: ScheduleService,
-        private readonly logger: LoggerService
+        private readonly logger: LoggerService,
+        private readonly http: HttpService
     ) {
         this.logger.setContext(ScheduleController.name);
     };
 
     @Post(CREATE_ROUTE)
     @ApiOperation({ summary: `Create a row of Schedule in postgres` })
-    async create(@Body() data: CreateScheduleDto): Promise<object> {
+    async create(@Body() data: CreateScheduleDto, @Res() response: Response, @Session() session: Record<string, any>): Promise<void | Response<any, Record<string, any>>> {
         try {
             this.logger.controllerDebug(CREATE_ROUTE);
-            return await this.scheduleService.create(data);
+            if (!session.visits) {
+                return response.status(401).redirect(REDIRECT_ROUTES);
+            } else {
+                await this.scheduleService.create(data);
+                return response.json(this.http.successResponse());
+            };
         } catch (err) {
             this.logger.errorMessage(err);
             throw new BadRequestException(err);
@@ -48,10 +57,15 @@ export class ScheduleController {
 
     @Get(READALL_ROUTE)
     @ApiOperation({ summary: 'Read all rows of Schedule in postgres' })
-    async readAll(): Promise<ScheduleSetup[]> {
+    async readAll(@Res() response: Response, @Session() session: Record<string, any>): Promise<void | Response<any, Record<string, any>>> {
         try {
             this.logger.controllerDebug(READALL_ROUTE);
-            return await this.scheduleService.readAll();
+            if (!session.visits) {
+                return response.status(401).redirect(REDIRECT_ROUTES);
+            } else {
+                const schedules = await this.scheduleService.readAll()
+                return response.json(schedules);
+            };
         } catch (err) {
             this.logger.errorMessage(err);
             throw new BadRequestException(err);
@@ -60,10 +74,15 @@ export class ScheduleController {
 
     @Patch(UPDATE_ROUTE)
     @ApiOperation({ summary: 'Update a row of Schedule in postgres by scheduleID' })
-    async update(@Body() data: UpdateScheduleDto): Promise<object> {
+    async update(@Body() data: UpdateScheduleDto, @Res() response: Response, @Session() session: Record<string, any>): Promise<void | Response<any, Record<string, any>>> {
         try {
             this.logger.controllerDebug(UPDATE_ROUTE);
-            return await this.scheduleService.update(data);
+            if (!session.visits) {
+                return response.status(401).redirect(REDIRECT_ROUTES);
+            } else {
+                await this.scheduleService.update(data);
+                return response.json(this.http.successResponse());
+            };
         } catch (err) {
             this.logger.errorMessage(err);
             throw new BadRequestException(err);
@@ -72,10 +91,15 @@ export class ScheduleController {
 
     @Delete(DELETE_ROUTE)
     @ApiOperation({ summary: 'Delete a row of Schedule in postgres by scheduleID' })
-    async delete(@Body() data: DeleteScheduleDto): Promise<object> {
+    async delete(@Body() data: DeleteScheduleDto, @Res() response: Response, @Session() session: Record<string, any>): Promise<void | Response<any, Record<string, any>>> {
         try {
             this.logger.controllerDebug(DELETE_ROUTE);
-            return await this.scheduleService.delete(data);
+            if (!session.visits) {
+                return response.status(401).redirect(REDIRECT_ROUTES);
+            } else {
+                await this.scheduleService.delete(data);
+                return response.json(this.http.successResponse());
+            };
         } catch (err) {
             this.logger.errorMessage(err);
             throw new BadRequestException(err);
